baseserver: maintain 'started' property rather than relying on subclasses to implement it.
With this change server.started always returns True after start() before stop() (or kill()).

Previously it could be False occassionally, because StreamServer could stop accepting for a while in presence of errors and StreamServer.started was defined as "whether the server is currently accepting".

Also, in StreamServer.kill remove link installed on pool's semaphore to avoid a GC cycle.