Semaphore: avoid copying list of links; rawlink() no longer schedules notification
Avoid copying list of links by maintaining a 'dirty' flag that causes
_notify_links to restart iteration. This makes sure that we don't miss
a recently added link because notification is already in progress.

It is possible to notify the same links (that did not remove themselvers)
more than once, but it's better than not notifying some links at all.

rawlink() no longer schedules notification immediatelly if needed, because
there seem to be utility in such behaviour.