fix issues with tracing and thread shutdown
python's threading module relies on the lock's class __exit__ to not
call into the tracing function set via sys.settrace until the lock
really is released. The only way to prevent this is to implement it in
C. So, we move coros.Semaphore to a cython module.

fixes the failing tests in test__threading_vs_settrace.py